requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
durationpy==0.9
favicon==0.7.0
Flask==3.1.0
google-auth==2.38.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kubernetes==32.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
pyasn1_modules==0.4.2
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests-oauthlib==2.0.0
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
